+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes to Calva.
 - [Add modifiers `pr-str`, `replace`, and `replace-first` to command snippet variables](https://github.com/BetterThanTomorrow/calva/issues/2282)
 - Fix: [Custom REPL Command/Snippet: `repl` setting not honored](https://github.com/BetterThanTomorrow/calva/issues/2770)
 - [Add `$ns-form` custom command variable](https://github.com/BetterThanTomorrow/calva/issues/2771)
+- Fix: [Custom REPL Commands are evaluated with REPL Window context](https://github.com/BetterThanTomorrow/calva/issues/2773)
 
 ## [2.0.496] - 2025-04-06